Split message utilities from overall message processing
